% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_mean}
\alias{focal_mean}
\title{Focal mean with a given kernel}
\usage{
focal_mean(x, kernel, global)
}
\arguments{
\item{x}{A matrix.}

\item{kernel}{A kernel which can be produced by \code{\link{kernel_gauss}}.}

\item{global}{Are data at global scale? If \code{TRUE}, both vertical
borders will be padded.}
}
\value{
A matrix of focal mean.
}
\description{
Focal mean with a given kernel
}
