% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_stat_tw}
\alias{focal_stat_tw}
\title{Focal statistics on temporal window}
\usage{
focal_stat_tw(x, y, fuzzy, xmin = NA_real_, xmax = NA_real_,
  ymin = NA_real_, ymax = NA_real_, stat = "xmn")
}
\arguments{
\item{x}{A 3d array. The 3rd one is time.}

\item{y}{The other array.}

\item{fuzzy}{Fuzzify or not prior to calculation.}

\item{xmin, xmax, ymin, ymax}{Parameters that determine fuzzy sets.
If \code{NA_real_}, the default, will be set as extrema of data.}

\item{stat}{Statistic to return. One of 'xmn', 'ymn', 'xsd', or 'ysd'.}
}
\value{
A matrix.
}
\description{
This function computes the GCSM at each cell treating the 3rd dimension
as a temporal window.
}
