% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_attr.R
\name{get_set_attr}
\alias{get_set_attr}
\alias{get_attr}
\alias{set_attr}
\title{get and set attribute of stars object}
\usage{
get_attr(x, attr)

set_attr(x, attr, ar_wide)
}
\arguments{
\item{x}{object of class \code{stars}}

\item{attr}{charter indicating attribute}

\item{ar_wide}{a wide-from matrix}
}
\value{
\code{get_attr} returns a wide-from matrix

\code{set_attr} returns a stars object
}
\description{
\code{get_attr} retrieves an attribute from stars, and rearranges to wide-form.
\code{set_attr}, on the contrary, puts a wide-form matrix to stars as an attribute.
}
\examples{
library(stars)
x <- matrix(1:18, nrow=6)
st <- x \%>\% st_as_stars() \%>\% setNames('v')
st \%>\% image(text_values = TRUE)
st \%>\% get_attr('v')
}
