% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_gauss}
\alias{kernel_gauss}
\title{Gaussian kernel}
\usage{
kernel_gauss(ksize, sigma)
}
\arguments{
\item{ksize}{Side length of the squared kernel.}

\item{sigma}{Standard deviation of Gaussian weighting function.}
}
\value{
A squared matrix sum to 1.
}
\description{
Gaussian kernel
}
