% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_stat_sw}
\alias{focal_stat_sw}
\title{Focal statistics on spatial window}
\usage{
focal_stat_sw(x, y, fuzzy, xmin = NA_real_, xmax = NA_real_,
  ymin = NA_real_, ymax = NA_real_, ksize = 9, global = FALSE,
  stat = "xmn")
}
\arguments{
\item{x}{A matrix.}

\item{y}{The other matrix.}

\item{fuzzy}{Fuzzify or not prior to calculation.}

\item{xmin}{Lower limit of \code{x}. Defaults to \code{min(x)}.}

\item{xmax}{Upper limit of \code{x}. Defaults to \code{max(x)}.}

\item{ymin}{Lower limit of \code{y}. Defaults to \code{min(y)}.}

\item{ymax}{Upper limit of \code{y}. Defaults to \code{max(y)}.}

\item{ksize}{Side length of the spatial window.}

\item{global}{Are data at global scale? If \code{TRUE}, both vertical
borders will be padded.}

\item{stat}{Statistic to return. One of \code{c('xmn', 'ymn', 'xsd', 'ysd')}.}
}
\value{
A matirx of \code{stat}.
}
\description{
Focal statistics on spatial window
}
