% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_stat_sw}
\alias{focal_stat_sw}
\title{Focal statistics on spatial window}
\usage{
focal_stat_sw(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  ksize = 9,
  globe = FALSE,
  stat = "xmn"
)
}
\arguments{
\item{x}{A matrix.}

\item{y}{The other matrix.}

\item{rescale}{Rescale or not before computation.}

\item{xmin, xmax, ymin, ymax}{Normalization parameters. If \code{NA}, are calculated
from the ranges of \code{x} and \code{y}, respectively.}

\item{ksize}{Side length of the spatial windows.}

\item{globe}{Are data at the global scale? If \code{TRUE}, two vertical borders
will be padded before computation.}

\item{stat}{Statistic to return. One of 'xmn', 'ymn', 'xsd', or 'ysd'.}
}
\value{
A matirx.
}
\description{
Focal statistics on spatial window
}
