% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_stat_tw}
\alias{focal_stat_tw}
\title{Focal statistics on temporal window}
\usage{
focal_stat_tw(
  xxx,
  yyy,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  stat = "xmn"
)
}
\arguments{
\item{xxx}{A 3-d array with the 3rd dimension representing time.}

\item{yyy}{The other 3-d array.}

\item{rescale}{Rescale or not before computation.}

\item{xmin, xmax, ymin, ymax}{Normalization parameters. If \code{NA}, are calculated
from the ranges of \code{xxx} and \code{yyy}, respectively.}

\item{stat}{Statistic to return. One of 'xmn', 'ymn', 'xsd', or 'ysd'.}
}
\value{
A matrix.
}
\description{
This function computes the GCSM at each cell treating the 3rd dimension
as a temporal window.
}
