% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_stat_tw}
\alias{focal_stat_tw}
\title{Focal statistics on temporal window}
\usage{
focal_stat_tw(x, y, fuzzy, xmin = NA_real_, xmax = NA_real_,
  ymin = NA_real_, ymax = NA_real_, stat = "xmn")
}
\arguments{
\item{x}{A 3d array. The 3rd one is time.}

\item{y}{The other array.}

\item{fuzzy}{Fuzzify or not prior to calculation.}

\item{xmin}{Lower limit of \code{x}. Defaults to `min(x).}

\item{xmax}{Upper limit of \code{x}. Defaults to \code{max(x)}.}

\item{ymin}{Lower limit of \code{y}. Defaults to `min(y).}

\item{ymax}{Upper limit of \code{y}. Defaults to `max(y).}

\item{stat}{Statistic to return. One of \code{c('xmn', 'ymn', 'xsd', 'ysd')}.}
}
\value{
A matrix of \code{stat}.
}
\description{
This function computes the GCSM at each cell treating the 3rd dimension
as a temporal window.
}
